requirements.txt
aiohttp==3.10.11
aiosignal==1.3.1
arcp==0.2.1
argo-workflows==6.5.8
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bioblend==1.3.0
CacheControl==0.14.0
cachetools==5.4.0
cattrs==24.1.2
certifi==2024.07.04
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.9.0
cryptography==43.0.1
Django==5.0.11
django-allauth==64.0.0
django-browser-reload==1.13.0
django-compat==1.0.15
django-environ==0.11.2
django-json-ld==0.0.5
django-static-fontawesome==6.5.2.0
django4-background-tasks==1.2.10
django_signposting==0.10.0
djangorestframework==3.15.2
djangorestframework-api-key==3.1.0
filelock==3.14.0
frozendict==2.4.4
frozenlist==1.4.1
galaxy2cwl==0.1.4
gunicorn==22.0.0
gxformat2==0.18.0
html5lib==1.1
httplink==0.2.0
idna==3.7
iniconfig==2.0.0
inquirerpy==0.3.4
isodate==0.6.1
Jinja2==3.1.5
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
owlrl==6.0.2
packaging==24.0
pfzy==0.3.4
platformdirs==4.3.6
pluggy==1.5.0
prettytable==3.12.0
prompt_toolkit==3.0.48
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyLD==2.0.4
pyparsing==3.1.2
pyshacl==0.26.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
python3-saml==1.16.0
PyYAML==6.0.1
rdflib==7.0.0
requests==2.32.3
requests-cache==1.2.1
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3987==1.3.8
rich==13.9.4
rich-click==1.8.4
roc-validator @ git+https://github.com/kikkomep/rocrate-validator.git@dc581d436776fc52cf3e458cb1fe2ffbd3c4c5ea
rocrate @ git+https://github.com/ResearchObject/ro-crate-py.git@b3b2438804db2d3adc279f18c5e392a22fa6b516
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
schema-salad==8.5.20240503091721
signposting==0.9.9
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
tinydb==4.8.0
toml==0.10.2
tuspy==1.0.3
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xmlsec==1.3.14
yarl==1.12.0
